+++ src/args.rs
@@ -1,18 +1,26 @@
+//! Parses command line arguments.
+
+use crate::{config::Config, Command, ErrorMessage};
 use std::path::{Path, PathBuf};
 use std::{env, mem};
-use Command;
 
-pub(crate) fn parse_args() -> Command {
-    let mut args = env::args().skip(1);
+pub(crate) fn parse_args() -> Result<Command, ErrorMessage> {
+    let mut args = env::args();
+    let executable_name = args.next().ok_or("no first argument (executable name)")?;
     let first = args.next();
     match first.as_ref().map(|s| s.as_str()) {
         Some("build") => parse_build_args(args),
-        Some("run") => match parse_build_args(args) {
+        Some("bootimage") if executable_name.ends_with("cargo-bootimage") => parse_build_args(args)
+            .map(|cmd| match cmd {
+                Command::BuildHelp => Command::CargoBootimageHelp,
+                cmd => cmd,
+            }),
+        Some("run") => parse_build_args(args).map(|cmd| match cmd {
             Command::Build(args) => Command::Run(args),
             Command::BuildHelp => Command::RunHelp,
             cmd => cmd,
-        },
-        Some("test") => match parse_build_args(args) {
+        }),
+        Some("test") => parse_build_args(args).map(|cmd| match cmd {
             Command::Build(args) => {
                 assert_eq!(
                     args.bin_name, None,
@@ -22,14 +30,15 @@ pub(crate) fn parse_args() -> Command {
             }
             Command::BuildHelp => Command::TestHelp,
             cmd => cmd,
-        },
-        Some("--help") | Some("-h") => Command::Help,
-        Some("--version") => Command::Version,
-        _ => Command::NoSubcommand,
+        }),
+        Some("runner") => parse_runner_args(args),
+        Some("--help") | Some("-h") => Ok(Command::Help),
+        Some("--version") => Ok(Command::Version),
+        _ => Ok(Command::NoSubcommand),
     }
 }
 
-fn parse_build_args<A>(args: A) -> Command
+fn parse_build_args<A>(args: A) -> Result<Command, ErrorMessage>
 where
     A: Iterator<Item = String>,
 {
@@ -40,13 +49,14 @@ where
     let mut cargo_args = Vec::new();
     let mut run_args = Vec::new();
     let mut run_args_started = false;
+    let mut quiet = false;
     {
-        fn set<T>(arg: &mut Option<T>, value: Option<T>) {
+        fn set<T>(arg: &mut Option<T>, value: Option<T>) -> Result<(), ErrorMessage> {
             let previous = mem::replace(arg, value);
-            assert!(
-                previous.is_none(),
-                "multiple arguments of same type provided"
-            )
+            if previous.is_some() {
+                Err("multiple arguments of same type provided")?
+            }
+            Ok(())
         };
 
         let mut arg_iter = args.into_iter();
@@ -57,14 +67,17 @@ where
             }
             match arg.as_ref() {
                 "--help" | "-h" => {
-                    return Command::BuildHelp;
+                    return Ok(Command::BuildHelp);
                 }
                 "--version" => {
-                    return Command::Version;
+                    return Ok(Command::Version);
+                }
+                "--quiet" => {
+                    quiet = true;
                 }
                 "--bin" => {
                     let next = arg_iter.next();
-                    set(&mut bin_name, next.clone());
+                    set(&mut bin_name, next.clone())?;
                     cargo_args.push(arg);
                     if let Some(next) = next {
                         cargo_args.push(next);
@@ -74,12 +87,12 @@ where
                     set(
                         &mut bin_name,
                         Some(String::from(arg.trim_start_matches("--bin="))),
-                    );
+                    )?;
                     cargo_args.push(arg);
                 }
                 "--target" => {
                     let next = arg_iter.next();
-                    set(&mut target, next.clone());
+                    set(&mut target, next.clone())?;
                     cargo_args.push(arg);
                     if let Some(next) = next {
                         cargo_args.push(next);
@@ -89,7 +102,7 @@ where
                     set(
                         &mut target,
                         Some(String::from(arg.trim_start_matches("--target="))),
-                    );
+                    )?;
                     cargo_args.push(arg);
                 }
                 "--manifest-path" => {
@@ -101,7 +114,7 @@ where
                                 .canonicalize()
                                 .expect("--manifest-path invalid")
                         }),
-                    );
+                    )?;
                     cargo_args.push(arg);
                     if let Some(next) = next {
                         cargo_args.push(next);
@@ -111,11 +124,11 @@ where
                     let path = Path::new(arg.trim_start_matches("--manifest-path="))
                         .canonicalize()
                         .expect("--manifest-path invalid");
-                    set(&mut manifest_path, Some(path));
+                    set(&mut manifest_path, Some(path))?;
                     cargo_args.push(arg);
                 }
                 "--release" => {
-                    set(&mut release, Some(true));
+                    set(&mut release, Some(true))?;
                     cargo_args.push(arg);
                 }
                 "--" => {
@@ -128,14 +141,15 @@ where
         }
     }
 
-    Command::Build(Args {
+    Ok(Command::Build(Args {
         cargo_args,
         run_args,
         bin_name,
         target,
         manifest_path,
         release: release.unwrap_or(false),
-    })
+        quiet,
+    }))
 }
 
 #[derive(Debug, Clone)]
@@ -144,6 +158,8 @@ pub struct Args {
     pub cargo_args: Vec<String>,
     /// All arguments that are passed to the runner.
     pub run_args: Vec<String>,
+    /// Suppress any output to stdout.
+    pub quiet: bool,
     /// The manifest path (also present in `cargo_args`).
     manifest_path: Option<PathBuf>,
     /// The name of the binary (passed `--bin` argument) (also present in `cargo_args`).
@@ -159,16 +175,8 @@ impl Args {
         &self.manifest_path
     }
 
-    pub fn bin_name(&self) -> &Option<String> {
-        &self.bin_name
-    }
-
     pub fn target(&self) -> &Option<String> {
         &self.target
-    }
-
-    pub fn release(&self) -> bool {
-        self.release
     }
 
     pub fn set_target(&mut self, target: String) {
@@ -184,4 +192,62 @@ impl Args {
         self.cargo_args.push("--bin".into());
         self.cargo_args.push(bin_name);
     }
+
+    pub fn apply_default_target(&mut self, config: &Config, crate_root: &Path) {
+        if self.target().is_none() {
+            if let Some(ref target) = config.default_target {
+                let canonicalized_target = crate_root.join(target);
+                self.set_target(canonicalized_target.to_string_lossy().into_owned());
+            }
+        }
+    }
+}
+
+fn parse_runner_args<A>(args: A) -> Result<Command, ErrorMessage>
+where
+    A: Iterator<Item = String>,
+{
+    let mut executable = None;
+    let mut quiet = false;
+
+    let mut arg_iter = args.into_iter().fuse();
+
+    loop {
+        match arg_iter.next().as_ref().map(|s| s.as_str()) {
+            Some("--help") | Some("-h") => {
+                return Ok(Command::RunnerHelp);
+            }
+            Some("--version") => {
+                return Ok(Command::Version);
+            }
+            Some("--quiet") => {
+                quiet = true;
+            }
+            Some(exe) if executable.is_none() => {
+                let path = Path::new(exe);
+                let path_canonicalized = path.canonicalize().map_err(|err| {
+                    format!(
+                        "Failed to canonicalize executable path `{}`: {}",
+                        path.display(),
+                        err
+                    )
+                })?;
+                executable = Some(path_canonicalized);
+            }
+            Some(arg) => Err(format!("unexpected argument `{}`", arg))?,
+            None => break,
+        }
+    }
+
+    Ok(Command::Runner(RunnerArgs {
+        executable: executable.ok_or("excepted path to kernel executable as first argument")?,
+        quiet,
+    }))
+}
+
+#[derive(Debug, Clone)]
+pub struct RunnerArgs {
+    pub executable: PathBuf,
+    /// Suppress any output to stdout.
+    pub quiet: bool,
 }